requirements.txt
slack-bolt>=1.21.0,<2.0
slack-sdk>=3.33.0,<4.0
gspread>=6.0.0,<7.0
google-auth>=2.0.0,<3.0
python-dotenv>=1.0.0,<2.0
pytest>=8.0.0,<9.0
